refactor node to include parent link
