+++ byte-rs/src/byte/nodes/mod.rs
@@ -1,210 +1,16 @@
 use super::*;
 
 pub mod eval;
+pub mod node_value;
 pub mod operators;
 pub mod parsing;
 
 pub use eval::*;
+pub use node_value::*;
 pub use operators::*;
 pub use parsing::*;
 
 const SHOW_INDENT: bool = false;
-
-/// Enumeration of all available language elements.
-///
-/// Nodes relate to the source code, representing language constructs of all
-/// levels, from files, raw text, and tokens, all the way to fully fledged
-/// definitions.
-#[derive(Clone, Debug, Eq, PartialEq, Hash)]
-pub enum NodeValue {
-	Token(Token),
-	Null,
-	Boolean(bool),
-	//----[ Structural ]------------------------------------------------------//
-	Raw(Vec<Node>),
-	Line(Node),
-	Sequence(Vec<Node>),
-	Group(Node),
-	Block(Node, Node),
-	//----[ Logic ]-----------------------------------------------------------//
-	If {
-		expr: Node,
-		if_true: Node,
-		if_false: Option<Node>,
-	},
-	For {
-		var: Symbol,
-		offset: usize,
-		from: Node,
-		to: Node,
-		body: Node,
-	},
-	//----[ AST ]-------------------------------------------------------------//
-	Let(Symbol, Option<usize>, Node),
-	UnaryOp(UnaryOp, Node),
-	BinaryOp(BinaryOp, Node, Node),
-	Variable(Symbol, Option<usize>),
-	Print(Node, &'static str),
-	Conditional(Node, Node, Node),
-}
-
-impl NodeValue {
-	pub fn at(self, scope: ScopeHandle, span: Span) -> Node {
-		Node::new(self, scope, at(span))
-	}
-
-	pub fn symbol(&self) -> Option<Symbol> {
-		let symbol = match self {
-			NodeValue::Token(Token::Word(symbol)) => symbol,
-			NodeValue::Token(Token::Symbol(symbol)) => symbol,
-			_ => return None,
-		};
-		Some(symbol.clone())
-	}
-
-	pub fn len(&self) -> usize {
-		self.children().len()
-	}
-
-	pub fn iter(&self) -> impl Iterator<Item = &Node> {
-		self.children().into_iter()
-	}
-
-	pub fn get(&self, index: usize) -> Option<&Node> {
-		self.children().get(index).cloned()
-	}
-
-	pub fn children(&self) -> Vec<&Node> {
-		match self {
-			NodeValue::Token(_) => vec![],
-			NodeValue::Null => vec![],
-			NodeValue::Boolean(_) => vec![],
-			NodeValue::Raw(ls) => ls.iter().map(|x| x).collect(),
-			NodeValue::Line(expr) => vec![expr],
-			NodeValue::Sequence(ls) => ls.iter().map(|x| x).collect(),
-			NodeValue::Group(it) => vec![it],
-			NodeValue::Block(head, body) => vec![head, body],
-			NodeValue::If {
-				expr,
-				if_true,
-				if_false,
-			} => {
-				let mut out = vec![expr, if_true];
-				if let Some(ref if_false) = if_false {
-					out.push(if_false);
-				}
-				out
-			}
-			NodeValue::For { from, to, body, .. } => vec![from, to, body],
-			NodeValue::Let(.., expr) => vec![expr],
-			NodeValue::UnaryOp(_, expr) => vec![expr],
-			NodeValue::BinaryOp(_, lhs, rhs) => vec![lhs, rhs],
-			NodeValue::Variable(..) => vec![],
-			NodeValue::Print(expr, _) => vec![expr],
-			NodeValue::Conditional(cond, t, f) => vec![cond, t, f],
-		}
-	}
-
-	pub fn get_dependencies<P: FnMut(&Node)>(&self, mut output: P) {
-		for it in self.children() {
-			output(it);
-		}
-	}
-}
-
-impl Display for NodeValue {
-	fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
-		let ctx = Context::get().format_without_span();
-		ctx.is_used();
-
-		match self {
-			NodeValue::Token(token) => write!(f, "{token}"),
-			NodeValue::Null => write!(f, "(null)"),
-			NodeValue::Boolean(value) => write!(f, "({value})"),
-			NodeValue::Raw(nodes) => match nodes.len() {
-				0 => write!(f, "<>"),
-				1 => {
-					let output = format!("{}", nodes[0]);
-					if output.contains('\n') || output.len() > 50 {
-						write!(f.indented(), "<\n{output}")?;
-						write!(f, "\n>")
-					} else {
-						write!(f, "<{output}>")
-					}
-				}
-				_ => {
-					let ctx = ctx.format_with_span();
-					ctx.is_used();
-					write!(f, "<# Raw:")?;
-					let mut out = f.indented();
-					for (n, it) in nodes.iter().enumerate() {
-						write!(out, "\n# {n}:")?;
-						write!(out.indented(), "\n{it}")?;
-					}
-					write!(f, "\n>")
-				}
-			},
-			NodeValue::Line(value) => {
-				let ctx = ctx.format_with_span();
-				ctx.is_used();
-				write!(f.indented(), "# Line:\n{value}")
-			}
-			NodeValue::Sequence(nodes) => {
-				let ctx = ctx.format_with_span();
-				ctx.is_used();
-				write!(f, "# Sequence:")?;
-				let mut out = f.indented();
-				for (n, it) in nodes.iter().enumerate() {
-					write!(out, "\n# {n}:")?;
-					write!(out.indented(), "\n{it}")?;
-				}
-				Ok(())
-			}
-			NodeValue::Group(value) => write!(f, "Group({value})"),
-			NodeValue::Block(head, body) => {
-				write!(f, "Block({head}:")?;
-				write!(f.indented(), "\n{body}")?;
-				write!(f, "\n)")
-			}
-			NodeValue::If {
-				expr,
-				if_true,
-				if_false,
-			} => {
-				write!(f, "if ({expr}) {{")?;
-				write!(f.indented(), "\n{if_true}")?;
-				if let Some(if_false) = if_false {
-					write!(f, "\n}} else {{")?;
-					write!(f.indented(), "\n{if_false}")?;
-				}
-				write!(f, "\n}}")
-			}
-			NodeValue::For {
-				var,
-				offset,
-				from,
-				to,
-				body,
-			} => {
-				write!(f, "for {var}#{offset} in {from}..{to} {{")?;
-				write!(f.indented(), "\n{body}")?;
-				write!(f, "\n}}")
-			}
-			NodeValue::Let(name, offset, expr) => {
-				let offset = offset.unwrap_or(0);
-				write!(f, "let {name}_{offset} = {expr}")
-			}
-			NodeValue::UnaryOp(op, arg) => write!(f, "{op} {arg}"),
-			NodeValue::BinaryOp(op, lhs, rhs) => write!(f, "({lhs} {op} {rhs})"),
-			NodeValue::Variable(var, offset) => {
-				let offset = offset.unwrap_or(0);
-				write!(f, "{var}_{offset}")
-			}
-			NodeValue::Print(expr, _) => write!(f, "Print({expr})"),
-			NodeValue::Conditional(cond, a, b) => write!(f, "{cond} ? {a} : {b}"),
-		}
-	}
-}
 
 #[derive(Clone)]
 pub struct Node {
@@ -217,6 +23,7 @@ struct NodeData {
 	value: RwLock<NodeValue>,
 	version: RwLock<usize>,
 	scope: ScopeHandle,
+	parent: RwLock<Option<(Weak<NodeData>, usize)>>,
 }
 
 impl Node {
@@ -225,21 +32,22 @@ impl Node {
 		let version = 0.into();
 		let id = id();
 		let span = span.into();
-		Self {
-			data: NodeData {
-				id,
-				span,
-				value,
-				version,
-				scope,
-			}
-			.into(),
-		}
+		let data = NodeData {
+			id,
+			span,
+			value,
+			version,
+			scope,
+			parent: Default::default(),
+		};
+		let node = Self { data: data.into() };
+		node.fixup_children(&node.data.value.read().unwrap());
+		node
 	}
 
 	pub fn raw(nodes: Vec<Node>, scope: ScopeHandle) -> Self {
 		let span = Span::from_node_vec(&nodes);
-		NodeValue::Raw(nodes).at(scope, span)
+		NodeValue::Raw(nodes.into()).at(scope, span)
 	}
 
 	pub fn id(&self) -> Id {
@@ -248,6 +56,38 @@ impl Node {
 
 	pub fn version(&self) -> usize {
 		*self.data.version.read().unwrap()
+	}
+
+	pub fn parent(&self) -> Option<Node> {
+		self.get_parent().map(|x| x.0)
+	}
+
+	pub fn next(&self) -> Option<Node> {
+		self.get_parent().and_then(|(node, index)| node.get(index + 1))
+	}
+
+	pub fn prev(&self) -> Option<Node> {
+		self.get_parent()
+			.and_then(|(node, index)| if index > 0 { node.get(index - 1) } else { None })
+	}
+
+	fn get_parent(&self) -> Option<(Node, usize)> {
+		let parent = self.data.parent.read().unwrap();
+		parent.as_ref().and_then(|(data, index)| {
+			if let Some(data) = data.upgrade() {
+				Some((Node { data }, *index))
+			} else {
+				None
+			}
+		})
+	}
+
+	fn set_parent(&self, new_parent: Option<(&Node, usize)>) {
+		let new_parent = new_parent.map(|(node, index)| (Arc::downgrade(&node.data), index));
+		self.write(|| {
+			let mut parent = self.data.parent.write().unwrap();
+			*parent = new_parent;
+		});
 	}
 
 	pub fn val(&self) -> NodeValue {
@@ -279,12 +119,30 @@ impl Node {
 	}
 
 	pub fn set_value(&mut self, new_value: NodeValue, new_span: Span) {
-		let mut value = self.data.value.write().unwrap();
-		let mut span = self.data.span.write().unwrap();
+		self.write(|| {
+			let mut value = self.data.value.write().unwrap();
+			let mut span = self.data.span.write().unwrap();
+
+			for it in value.children() {
+				it.set_parent(None);
+			}
+			self.fixup_children(&new_value);
+
+			*value = new_value;
+			*span = new_span;
+		});
+	}
+
+	fn write<T, P: FnOnce() -> T>(&self, write: P) -> T {
 		let mut version = self.data.version.write().unwrap();
-		*value = new_value;
-		*span = new_span;
 		*version = *version + 1;
+		(write)()
+	}
+
+	fn fixup_children(&self, value: &NodeValue) {
+		for (index, it) in value.children().iter().enumerate() {
+			it.set_parent(Some((self, index)));
+		}
 	}
 
 	//----------------------------------------------------------------------------------------------------------------//
@@ -314,7 +172,8 @@ impl Node {
 		let slice = &list[range];
 		let span = Span::from_nodes(slice);
 		let span = span.or_with(|| list.get(index).map(|x| x.span().pos()).unwrap_or_default());
-		NodeValue::Raw(slice.iter().map(|x| (*x).clone()).collect()).at(scope, span)
+		let list = Arc::new(slice.iter().map(|x| (*x).clone()).collect());
+		NodeValue::Raw(list).at(scope, span)
 	}
 
 	/// Iterator over this node's children.
@@ -331,7 +190,7 @@ impl Node {
 	pub fn to_raw(self) -> Node {
 		let span = self.span();
 		let scope = self.scope_handle();
-		NodeValue::Raw(vec![self]).at(scope, span)
+		NodeValue::Raw(vec![self].into()).at(scope, span)
 	}
 
 	pub fn is_symbol(&self, expected: &Symbol) -> bool {
@@ -357,6 +216,30 @@ impl Node {
 
 	pub fn symbol(&self) -> Option<Symbol> {
 		self.val().symbol()
+	}
+
+	pub(crate) fn sanity_check(&self) {
+		let mut prev = None;
+		let mut next = self.get(0);
+		for it in self.val().children() {
+			assert_eq!(Some(it), next.as_ref());
+			next = it.next();
+			let parent = it.parent();
+			if parent.as_ref().map(|x| x.id()) != Some(self.id()) {
+				let msg = "\n\nsanity check failed: child/parent is broken:\n\n";
+				panic!(
+					"{msg}-> child:\n{it}\n\n-> should be parent:\n{self}\n\n-> but was: {}\n\n",
+					if let Some(parent) = parent {
+						format!("{parent}")
+					} else {
+						format!("(none)")
+					}
+				);
+			}
+			assert!(it.prev() == prev);
+			prev = Some(it.clone());
+			it.sanity_check();
+		}
 	}
 }
