+++ byte-rs/src/byte/nodes/operators/op_comma.rs
@@ -13,7 +13,7 @@ impl ParseSplitSequence for CommaOperator {
 
 	fn new_node(&self, ctx: &mut EvalContext, node: Vec<Node>, span: Span) -> Result<Node> {
 		let _ = ctx;
-		Ok(NodeValue::Sequence(node).at(ctx.scope_handle(), span))
+		Ok(NodeValue::Sequence(node.into()).at(ctx.scope_handle(), span))
 	}
 }
